requirements.txt
google-cloud-firestore==2.3.0
google-auth==2.6.0
google-api-python-client==2.31.0
requests==2.26.0
numpy==1.21.2
pandas==1.3.3
tensorflow==2.6.0
transformers==4.11.3